Final biome changes
